Extract Jörmungandr Client in its own module
