+++ lib/jormungandr/src/Cardano/Wallet/Jormungandr/Network.hs
@@ -9,7 +9,6 @@
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TupleSections #-}
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 
 -- |
@@ -18,7 +17,7 @@
 --
 -- This module allows the wallet to retrieve blocks from a known @Jormungandr@
 -- node. This is done by providing a @NetworkLayer@ with some logic building on
--- top of an underlying @JormungandrLayer@ HTTP client.
+-- top of an underlying @JormungandrClient@ HTTP client.
 --
 -- It also provides facilities for starting the @jormungandr@ node backend
 -- process. The Jormungandr 'NetworkLayer' implementation uses the HTTP REST API
@@ -33,23 +32,23 @@ module Cardano.Wallet.Jormungandr.Network
 
     -- * Launching the node backend
     , JormungandrConfig (..)
-    , JormungandrLayer (..)
     , withJormungandr
     , connParamsPort
 
     -- * Errors
+    , ErrGetBlock (..)
     , ErrGetBlockchainParams (..)
     , ErrGetDescendants (..)
+    , ErrNetworkTip (..)
+    , ErrNetworkUnavailable (..)
+    , ErrPostTx (..)
     , ErrStartup (..)
     , ErrUnexpectedNetworkFailure (..)
 
-    -- * Re-export
-    , BaseUrl (..)
-    , Scheme (..)
-
     -- * Internal constructors
     , mkRawNetworkLayer
-    , mkJormungandrLayer
+    , BaseUrl (..)
+    , Scheme (..)
     ) where
 
 import Prelude
@@ -65,18 +64,29 @@ import Cardano.Launcher
     , transformLauncherTrace
     , withBackendProcess
     )
-import Cardano.Wallet.Jormungandr.Api
-    ( BlockId (..)
-    , GetBlock
-    , GetBlockDescendantIds
-    , GetStakeDistribution
-    , GetTipId
-    , PostMessage
-    , StakeApiResponse
-    , api
+import Cardano.Wallet.Jormungandr.Api.Client
+    ( BaseUrl (..)
+    , ErrGetBlock (..)
+    , ErrGetBlockchainParams (..)
+    , ErrGetDescendants (..)
+    , ErrNetworkTip (..)
+    , ErrNetworkUnavailable (..)
+    , ErrPostTx (..)
+    , ErrUnexpectedNetworkFailure (..)
+    , JormungandrClient
+    , LiftError (..)
+    , Scheme (..)
+    , defaultManagerSettings
+    , getBlockHeader
+    , getBlocks
+    , getInitialBlockchainParameters
+    , getTipId
+    , mkJormungandrClient
+    , newManager
+    , postMessage
     )
 import Cardano.Wallet.Jormungandr.Binary
-    ( ConfigParam (..), Message (..), convertBlock, runGetOrFail )
+    ( convertBlock, runGetOrFail )
 import Cardano.Wallet.Jormungandr.BlockHeaders
     ( BlockHeaders (..)
     , appendBlockHeaders
@@ -87,41 +97,21 @@ import Cardano.Wallet.Jormungandr.BlockHeaders
     , updateUnstableBlocks
     )
 import Cardano.Wallet.Jormungandr.Compatibility
-    ( Jormungandr, genConfigFile, localhostBaseUrl, softTxMaxSize )
+    ( Jormungandr, genConfigFile, localhostBaseUrl )
 import Cardano.Wallet.Jormungandr.Primitive.Types
     ( Tx )
 import Cardano.Wallet.Network
-    ( Cursor
-    , ErrGetBlock (..)
-    , ErrNetworkTip (..)
-    , ErrNetworkUnavailable (..)
-    , ErrPostTx (..)
-    , NetworkLayer (..)
-    , NextBlocksResult (..)
-    , defaultRetryPolicy
-    )
+    ( Cursor, NetworkLayer (..), NextBlocksResult (..), defaultRetryPolicy )
 import Cardano.Wallet.Network.Ports
     ( PortNumber, getRandomPort, waitForPort )
 import Cardano.Wallet.Primitive.Model
     ( BlockchainParameters (..) )
 import Cardano.Wallet.Primitive.Types
-    ( Block (..)
-    , BlockHeader (..)
-    , Hash (..)
-    , SlotId (..)
-    , SlotLength (..)
-    , TxWitness (..)
-    )
-import Control.Arrow
-    ( left )
+    ( Block (..), BlockHeader (..), Hash (..), SlotId (..) )
 import Control.Concurrent.MVar.Lifted
     ( MVar, modifyMVar, newMVar, readMVar )
 import Control.Exception
     ( Exception, bracket )
-import Control.Monad
-    ( void )
-import Control.Monad.Catch
-    ( throwM )
 import Control.Monad.IO.Class
     ( liftIO )
 import Control.Monad.Trans.Class
@@ -134,38 +124,12 @@ import Data.Coerce
     ( coerce )
 import Data.Function
     ( (&) )
-import Data.Maybe
-    ( mapMaybe )
-import Data.Proxy
-    ( Proxy (..) )
 import Data.Quantity
     ( Quantity (..) )
 import Data.Text
     ( Text )
 import Data.Word
     ( Word32 )
-import Network.HTTP.Client
-    ( Manager, defaultManagerSettings, newManager )
-import Network.HTTP.Types.Status
-    ( status400 )
-import Safe
-    ( tailSafe )
-import Servant.API
-    ( (:<|>) (..) )
-import Servant.Client
-    ( BaseUrl (..)
-    , ClientM
-    , Scheme (..)
-    , client
-    , mkClientEnv
-    , responseBody
-    , responseStatusCode
-    , runClientM
-    )
-import Servant.Client.Core
-    ( ServantError (..) )
-import Servant.Links
-    ( Link, safeLink )
 import System.Directory
     ( removeFile )
 import System.FilePath
@@ -175,7 +139,6 @@ import qualified Cardano.Wallet.Jormungandr.Binary as J
 import qualified Data.ByteString.Lazy as BL
 import qualified Data.Char as C
 import qualified Data.Text as T
-import qualified Data.Text.Encoding as T
 import qualified Data.Yaml as Yaml
 
 -- | Whether to start Jormungandr with the given config, or to connect to an
@@ -255,7 +218,7 @@ newNetworkLayer
 newNetworkLayer baseUrl block0H = do
     mgr <- liftIO $ newManager defaultManagerSettings
     st <- newMVar emptyBlockHeaders
-    let jor = mkJormungandrLayer mgr baseUrl
+    let jor = mkJormungandrClient mgr baseUrl
     g0 <- getInitialBlockchainParameters jor (coerce block0H)
     return (convertBlock <$> mkRawNetworkLayer g0 st jor)
 
@@ -271,7 +234,7 @@ mkRawNetworkLayer
         )
     => (block, BlockchainParameters)
     -> MVar BlockHeaders
-    -> JormungandrLayer m
+    -> JormungandrClient m
     -> NetworkLayer m t block
 mkRawNetworkLayer (block0, bp) st j = NetworkLayer
     { networkTip =
@@ -487,231 +450,6 @@ parseBlock0H file = parse <$> BL.readFile file
 connParamsPort :: JormungandrConnParams -> Int
 connParamsPort (JormungandrConnParams _ url) = baseUrlPort url
 
-{-------------------------------------------------------------------------------
-                            Jormungandr Client
--------------------------------------------------------------------------------}
-
--- | Endpoints of the jormungandr REST API.
-data JormungandrLayer m = JormungandrLayer
-    { getTipId
-        :: ExceptT ErrNetworkUnavailable m (Hash "BlockHeader")
-    , getBlock
-        :: Hash "BlockHeader"
-        -> ExceptT ErrGetBlock m J.Block
-    , getDescendantIds
-        :: Hash "BlockHeader"
-        -> Word
-        -> ExceptT ErrGetDescendants m [Hash "BlockHeader"]
-    , postMessage
-        :: (Tx, [TxWitness])
-        -> ExceptT ErrPostTx m ()
-    , getInitialBlockchainParameters
-        :: Hash "Genesis"
-        -> ExceptT ErrGetBlockchainParams m (J.Block, BlockchainParameters)
-    , getStakeDistribution
-        :: ExceptT ErrNetworkUnavailable m StakeApiResponse
-    }
-
--- Fetch a batch of blocks after but not including the given header.
-getBlocks
-    :: Monad m
-    => JormungandrLayer m
-    -> BlockHeader
-    -> ExceptT ErrGetBlock m [(Hash "BlockHeader", J.Block)]
-getBlocks j tip = do
-    -- Get the descendants of the tip's /parent/.
-    -- The first descendant is therefore the current tip itself. We need to
-    -- skip it. Hence the 'tail'.
-    ids <- withExceptT liftE $
-        tailSafe <$> getDescendantIds j (prevBlockHash tip) batchSize
-    mapM (\blockId -> (blockId,) <$> getBlock j blockId) ids
-  where
-    -- Maximum number of blocks to return from nextBlocks.
-    batchSize = 1000
-
--- | Get a block header corresponding to a header hash.
-getBlockHeader
-    :: Monad m
-    => JormungandrLayer m
-    -> Hash "BlockHeader"
-    -> ExceptT ErrGetBlock m (BlockHeader, Quantity "block" Word32)
-getBlockHeader j t = do
-    blk@(J.Block blkHeader _) <- getBlock j t
-    let nodeHeight = Quantity $ fromIntegral $ J.chainLength blkHeader
-    pure (header (convertBlock blk), nodeHeight)
-
--- | Construct a 'JormungandrLayer'-client
---
--- >>> mgr <- newManager defaultManagerSettings
--- >>> j = mkJormungandrLayer mgr (BaseUrl Http "localhost" 8080 "")
---
--- >>> (Right tip) <- runExceptT $ getTipId j
--- >>> tip
--- BlockId (Hash {getHash = "26c640a3de09b74398c14ca0a137ec78"})
---
--- >>> (Right block) <- runExceptT $ getBlock j t
--- >>> block
--- >>> Block {header = BlockHeader {slotId = SlotId {epochNumber = 0, slotNumber = 0}, prevBlockHash = Hash {getHash = "\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL\NUL"}}, transactions = [Tx {inputs = [], outputs = [TxOut {address = Address {unAddress = "3$\195xi\193\"h\154\&5\145}\245:O\"\148\163\165/h^\ENQ\245\248\229;\135\231\234E/"}, coin = Coin {getCoin = 14}}]}]}
---
--- At the time of writing, we only have the genesis-block, but we should be
--- able to get its descendants.
---
--- >>> let genesisHash = BlockId (Hash {getHash = "&\198@\163\222\t\183C\152\193L\160\161\&7\236x\245\229\EOT\175\177\167\131\190\b\b/\174\212\177:\179"})
--- >>> runExceptT $ getDescendantIds j t 4
--- Right []
-mkJormungandrLayer
-    :: Manager -> BaseUrl -> JormungandrLayer IO
-mkJormungandrLayer mgr baseUrl = JormungandrLayer
-    { getTipId = ExceptT $ do
-        let ctx = safeLink api (Proxy @GetTipId)
-        run (getBlockId <$> cGetTipId) >>= defaultHandler ctx
-
-    , getBlock = \blockId -> ExceptT $ do
-        run (cGetBlock (BlockId blockId)) >>= \case
-            Left (FailureResponse e) | responseStatusCode e == status400 ->
-                return . Left . ErrGetBlockNotFound $ blockId
-            x -> do
-                let ctx = safeLink api (Proxy @GetBlock) (BlockId blockId)
-                left ErrGetBlockNetworkUnreachable <$> defaultHandler ctx x
-
-    , getDescendantIds = \parentId count -> ExceptT $ do
-        run (map getBlockId <$> cGetBlockDescendantIds (BlockId parentId) (Just count))  >>= \case
-            Left (FailureResponse e) | responseStatusCode e == status400 ->
-                return . Left $ ErrGetDescendantsParentNotFound parentId
-            x -> do
-                let ctx = safeLink
-                        api
-                        (Proxy @GetBlockDescendantIds)
-                        (BlockId parentId)
-                        (Just count)
-                left ErrGetDescendantsNetworkUnreachable <$> defaultHandler ctx x
-
-    -- Never returns 'Left ErrPostTxProtocolFailure'. Will currently return
-    -- 'Right ()' when submitting correctly formatted, but invalid transactions.
-    --
-    -- https://github.com/input-output-hk/jormungandr/blob/fe638a36d4be64e0c4b360ba1c041e8fa10ea024/jormungandr/src/rest/v0/message/post.rs#L25-L39
-    , postMessage = \tx -> void $ ExceptT $ do
-        run (cPostMessage tx) >>= \case
-            Left (FailureResponse e)
-                | responseStatusCode e == status400 -> do
-                    let msg = T.decodeUtf8 $ BL.toStrict $ responseBody e
-                    return $ Left $ ErrPostTxBadRequest msg
-            x -> do
-                let ctx = safeLink api (Proxy @PostMessage)
-                left ErrPostTxNetworkUnreachable <$> defaultHandler ctx x
-
-    , getInitialBlockchainParameters = \block0 -> do
-        jblock@(J.Block _ msgs) <- ExceptT $ run (cGetBlock (BlockId $ coerce block0)) >>= \case
-            Left (FailureResponse e) | responseStatusCode e == status400 ->
-                return . Left . ErrGetBlockchainParamsGenesisNotFound $ block0
-            x -> do
-                let ctx = safeLink api (Proxy @GetBlock) (BlockId $ coerce block0)
-                let networkUnreachable = ErrGetBlockchainParamsNetworkUnreachable
-                left networkUnreachable <$> defaultHandler ctx x
-
-        let params = mconcat $ mapMaybe getConfigParameters msgs
-              where
-                getConfigParameters = \case
-                    Initial xs -> Just xs
-                    _ -> Nothing
-
-        let mpolicy = mapMaybe getPolicy params
-              where
-                getPolicy = \case
-                    ConfigLinearFee x -> Just x
-                    _ -> Nothing
-
-        let mduration = mapMaybe getSlotDuration params
-              where
-                getSlotDuration = \case
-                    SlotDuration x -> Just x
-                    _ -> Nothing
-
-        let mblock0Date = mapMaybe getBlock0Date params
-              where
-                getBlock0Date = \case
-                    Block0Date x -> Just x
-                    _ -> Nothing
-
-        let mepochLength = mapMaybe getSlotsPerEpoch params
-              where
-                getSlotsPerEpoch = \case
-                    SlotsPerEpoch x -> Just x
-                    _ -> Nothing
-
-        let mStability = mapMaybe getStability params
-              where
-                getStability = \case
-                   EpochStabilityDepth x -> Just x
-                   _ -> Nothing
-
-        case (mpolicy, mduration, mblock0Date, mepochLength, mStability) of
-            ([policy],[duration],[block0Date], [epochLength], [stability]) ->
-                return
-                    ( jblock
-                    , BlockchainParameters
-                        { getGenesisBlockHash = block0
-                        , getGenesisBlockDate = block0Date
-                        , getFeePolicy = policy
-                        , getEpochLength = epochLength
-                        , getSlotLength = SlotLength duration
-                        , getTxMaxSize = softTxMaxSize
-                        , getEpochStability = stability
-                        }
-                    )
-            _ ->
-                throwE $ ErrGetBlockchainParamsIncompleteParams params
-
-    , getStakeDistribution = ExceptT $ do
-        let ctx = safeLink api (Proxy @GetStakeDistribution)
-        run cGetStakeDistribution >>= defaultHandler ctx
-    }
-  where
-    run :: ClientM a -> IO (Either ServantError a)
-    run query = runClientM query (mkClientEnv mgr baseUrl)
-
-    defaultHandler
-        :: Link
-        -> Either ServantError a
-        -> IO (Either ErrNetworkUnavailable a)
-    defaultHandler ctx = \case
-        Right c -> return $ Right c
-
-        -- The node has not started yet or has exited.
-        -- This could be recovered from by either waiting for the node
-        -- initialise, or restarting the node.
-        Left (ConnectionError e) ->
-            return $ Left $ ErrNetworkUnreachable e
-
-        -- Other errors (status code, decode failure, invalid content type
-        -- headers). These are considered to be programming errors, so crash.
-        Left e -> do
-            throwM (ErrUnexpectedNetworkFailure ctx e)
-
-    cGetTipId
-        :<|> cGetBlock
-        :<|> cGetBlockDescendantIds
-        :<|> cPostMessage
-        :<|> cGetStakeDistribution
-        = client api
-
-data ErrUnexpectedNetworkFailure
-    = ErrUnexpectedNetworkFailure Link ServantError
-    deriving (Show)
-
-instance Exception ErrUnexpectedNetworkFailure
-
-data ErrGetDescendants
-    = ErrGetDescendantsNetworkUnreachable ErrNetworkUnavailable
-    | ErrGetDescendantsParentNotFound (Hash "BlockHeader")
-    deriving (Show, Eq)
-
-data ErrGetBlockchainParams
-    = ErrGetBlockchainParamsNetworkUnreachable ErrNetworkUnavailable
-    | ErrGetBlockchainParamsGenesisNotFound (Hash "Genesis")
-    | ErrGetBlockchainParamsIncompleteParams [ConfigParam]
-    deriving (Show, Eq)
-
 data ErrStartup
     = ErrStartupNodeNotListening
     | ErrStartupGetBlockchainParameters ErrGetBlockchainParams
@@ -720,23 +458,3 @@ data ErrStartup
     deriving (Show, Eq)
 
 instance Exception ErrStartup
-
--- | Safely lift error into a bigger type.
-class LiftError lift where
-    liftE :: lift
-
-instance LiftError (ErrGetBlock -> ErrNetworkTip) where
-    liftE = \case
-        ErrGetBlockNotFound _ -> ErrNetworkTipNotFound
-        ErrGetBlockNetworkUnreachable e -> ErrNetworkTipNetworkUnreachable e
-
-instance LiftError (ErrNetworkUnavailable -> ErrGetBlock) where
-    liftE =
-        ErrGetBlockNetworkUnreachable
-
-instance LiftError (ErrGetDescendants -> ErrGetBlock) where
-    liftE = \case
-        ErrGetDescendantsNetworkUnreachable e ->
-            ErrGetBlockNetworkUnreachable e
-        ErrGetDescendantsParentNotFound tip ->
-            ErrGetBlockNotFound tip
